requirements.txt
Django==1.7.1
pyjade==3.0.0
six==1.8.0
sorl-thumbnail==11.12
wsgiref==0.1.2
